Changes
.fla --
changed arrow color to white so it can be seen better.

.swf --
new version.

Engine --
added a maximum velocity for shooting.
changed collision function so arrow can be checked.
changed what happens when you hit the left side of the screen to go back
to the end of the last lvl.

View --
added an option to change lvl to the first lvl so that player can go
back there.

Arrow --
just commented out the trace of rotation to clean up output.