Arrow Shooting
Arrows should shoot from top of screen.
I will change it to the player's location shortly.
COULD NOT TEST
I am getting an error from Flash telling me that I can't create a swf
because the folder is Read-Only.  I am going to restart my computer in
an attempt to fix this.