Change from 2.0 to 2.1
I changed everything over to version 2.1 of Box2DFlash.

If it is not working for you, double check that you are pointing to the
right library files in Flash AND FlashDevelop or whatever you are using.

Most things are the same, just everything is different.  Mostly watch
out for creating shapes/fixtures.