Grappling Hook 0.3a
Grappling hook works now. I found and fixed the main bug with it.

Now the only funny thing is that the rope is solid, so if you make one
when you are above it, you can balance on it.  That and there is a small
bug with joints not getting removed sometimes.